Done with the application.
I made revisions on all of my essays (some minor, some major), and
I added a bib file, because I made some egregious changes to the
real bib files (to save space) that I didn't want to commit.
But I wanted to be able to remake these applications just as they are.